refactor tests with deployment helper and correct errors
